Move hive base entity to separate module (#126095)
